Update phase2 test script to run with bit witdh 32

+++ experiments/test_phase2_aws.py
@@ -9,6 +9,7 @@ import time
 import logging
 import numpy as np
 import sys
+import itertools
 from math import sqrt, pow
 
 REMOTE_USER = 'ubuntu'
@@ -23,29 +24,41 @@ logging.basicConfig(format='%(asctime)s %(message)s')
 logger.setLevel(logging.INFO)
 
 
-def update_and_compile(ip, remote_ip):
+def update_and_compile(ip, remote_ip, bit_width=64, compile_oblivc=True):
     key = paramiko.RSAKey.from_private_key_file(KEY_FILE)
     client = paramiko.SSHClient()
     client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
     logger.info('Connecting to {0}:'.format(ip))
     client.connect(hostname=ip, username=REMOTE_USER, pkey=key)
 
-    cmd_compile = 'cd obliv-c; make clean; make RELEASE=1; cd ..; ' +\
+    cmd_compile_oblivc = 'cd obliv-c; make clean; make RELEASE=1; cd ..; ' +\
         'cd secure-distributed-linear-regression; ' +\
-        'git stash; git checkout master; git pull; make clean; ' +\
+        'git stash; git checkout optimized_cgd; git pull; make clean; ' +\
         'git submodule update --init; ' +\
         'cd lib/absentminded-crypto-kit/; ' +\
         'make OBLIVC_PATH=$(cd ../../../obliv-c && pwd); ' +\
-        'cd ../..; ' +\
-        'make OBLIVC_PATH=$(cd ../obliv-c && pwd) ' + \
-        'REMOTE_HOST={0} bin/test_linear_system; '.format(remote_ip) + \
-        'killall -9 test_linear_system'
+        'cd ../..; '
 
-    cmd_dont_compile = 'killall -9 test_linear_system'
-    cmd = cmd_compile if COMPILE else cmd_dont_compile 
-    logger.info('Compiling in {0}:'.format(ip))
-    logger.info('{0}'.format(cmd))
-    stdin, stdout, stderr = client.exec_command(cmd)
+    cmd_compile_phase2 = 'make OBLIVC_PATH=$(cd ../obliv-c && pwd) ' + \
+        'BIT_WIDTH_32={} '.format(1 if bit_width == 32 else 0) + \
+        'REMOTE_HOST={0} bin/test_linear_system; '.format(remote_ip) + \
+        'killall -9 test_linear_system' if COMPILE\
+            else 'killall -9 test_linear_system'
+
+    if compile_oblivc:
+        logger.info(
+            'Compiling obliv-c and absentminded-crypto-kit in {0}:'.format(ip))
+        logger.info('{0}'.format(cmd_compile_oblivc))
+        stdin, stdout, stderr = client.exec_command(cmd_compile_oblivc)
+        for line in stdout:
+            logger.info('... ' + line.strip('\n'))
+        for line in stderr:
+            logger.error('... ' + line.strip('\n'))
+
+    logger.info(
+        'Compiling secure linear system solving in {0}:'.format(ip))
+    logger.info('{0}'.format(cmd_compile_phase2))
+    stdin, stdout, stderr = client.exec_command(cmd_compile_phase2)
     for line in stdout:
         logger.info('... ' + line.strip('\n'))
     for line in stderr:
@@ -134,12 +147,16 @@ def parse_output(n, d, X, y, lambda_, alg, solution, condition_number,
             f.write('\n{0} {1} {2} {3} {4} {5} {6}'.format(n, d,
                 alg, ot_time, time, error, gate_count))
             if alg == 'cgd':
-                gate_count_after_iters =\
-                    gate_count - cgd_iter_gate_sizes[-1]
+                try:
+                    gate_count_after_iters =\
+                        gate_count - cgd_iter_gate_sizes[-1]
+                except IndexError:
+                    gate_count_after_iters = 0
                 cgd_iter_gate_sizes = map(
                     lambda x: x + gate_count_after_iters,
                     cgd_iter_gate_sizes)
-                assert len(cgd_iter_gate_sizes) == len(
+                assert not cgd_iter_gate_sizes or\
+                    len(cgd_iter_gate_sizes) == len(
                     cgd_iter_solutions), '{0}\n{1}\n{2}\n{3}'.format(
                         len(cgd_iter_solutions),
                         cgd_iter_solutions,
@@ -153,7 +170,7 @@ def parse_output(n, d, X, y, lambda_, alg, solution, condition_number,
                         i + 1, error_i,
                         cgd_iter_objective_values[i],
                         cgd_iter_times[i],
-                        cgd_iter_gate_sizes[i],
+                        cgd_iter_gate_sizes[i] if cgd_iter_gate_sizes else -1,
                         ot_time))
             f.write('\nsolution:')
             f.write('\n{0}'.format(d))
@@ -285,8 +302,8 @@ def generate_benchmark(dest_folder):
         sigma = 0.1
         logger.info('Generating instance: n = {0}, d = {1}'.format(n, d))
         tmp_file = tempfile.NamedTemporaryFile()
-	tmp_filepath = tmp_file.name
-	tmp_file.close()
+        tmp_filepath = tmp_file.name
+        tmp_file.close()
 
         (_, _, X, y, lambda_, beta, condition_number, objective_value) = \
             generate_lin_system_from_regression_problem(
@@ -303,11 +320,15 @@ def generate_benchmark(dest_folder):
                 filepath_in = os.path.join(
                     dest_folder, filename_in)
                 os.system('mv {0} {1}'.format(tmp_filepath, filepath_in))
-                instances.append((n, d, X, y, lambda_, beta, condition_number, objective_value, filepath_in))
+                instances.append(
+                    (n, d, X, y, lambda_, beta, condition_number,
+                        objective_value, filepath_in))
 
         done = all([count[i] == 3 for i in range(1, 11)])
-    logger.info('Done generating instances: {}'.format(map(lambda x: x[8], instances)))
+    logger.info(
+        'Done generating instances: {}'.format(map(lambda x: x[8], instances)))
     return instances
+
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='Runs phase 2 experiments. '
@@ -324,17 +345,17 @@ if __name__ == "__main__":
     parser.add_argument(
         '--run_accuracy_tests', action='store_true', help='Run accuracy tests')
     exec_file = 'bin/test_linear_system'
-    dest_folder = 'test/experiments/phase2_d_plots/'
+    dest_folder = 'experiments/phase2_time_new_cgd/'
     assert os.path.exists(dest_folder), '{0} does not exist.'.format(
         dest_folder)
-    #assert not os.listdir(dest_folder), '{0} is not empty.'.format(
-    #    dest_folder)
+    assert not os.listdir(dest_folder), '{0} is not empty.'.format(
+        dest_folder)
     args = parser.parse_args()
     RUN_LOCALLY = args.run_locally
 
     if args.run_accuracy_tests:
         assert RUN_LOCALLY, 'Accuracy tests can only be run locally.'
-        num_iters_cgd = 15
+        num_iters_cgd = 30
         instances = generate_benchmark(dest_folder)
         for (n, d, X, y, lambda_, beta, condition_number, objective_value, filepath_in) in instances:
             for alg in ['cgd', 'cholesky']:
@@ -362,78 +383,107 @@ if __name__ == "__main__":
                         out_filename, filepath_exec)
         sys.exit()
 
-    num_iters_cgd = 15
-    num_examples = 1
+    num_iters_cgd = 20
+    num_examples = 3
+
+    algs = ['cgd', 'cholesky']
+    sigmas = [0.1]
+    ds = [10, 20, 50, 100, 200, 500]
+    ns = [100000]
+    bit_widths = [32, 64]
+    current_bit_width = bit_widths[0]
+    confs = itertools.product(
+        range(num_examples),
+        algs,
+        sigmas,
+        ds,
+        ns,
+        bit_widths)
 
     ips = [args.remote_ip_1, args.remote_ip_2]
     if not RUN_LOCALLY:
         update_and_compile(ips[0], ips[1])
         update_and_compile(ips[1], ips[0])
 
-  
-    for i in range(num_examples):
-        for alg in ['cgd', 'cholesky']:
-            for sigma in [0.1]:
-                for d in [10, 20, 50, 100, 200, 500]:
-                    for n in [100000]:
-                        logger.info(
-                            'Running instance: n={0}, d={1}, sigma={2}, alg={3}, num_iters_cgd={4} run={5}'.
-                            format(n, d, sigma, alg, num_iters_cgd, i + args.instance_num_offset))
-                        filename_in = 'test_LS_{0}x{1}_{2}_{3}.in'.format(
-                            n, d, sigma, i + args.instance_num_offset)
-                        filepath_in = os.path.join(
-                            dest_folder, filename_in)
+    for (i, alg, sigma, d, n, bit_width) in confs:
+        # for i in range(num_examples):
+        #    for alg in ['cgd', 'cholesky']:
+        #        for sigma in [0.1]:
+        #            for d in [10, 20, 50, 100, 200, 500]:
+        #                for n in [100000]:
 
-                        (A, b, X, y, lambda_, beta, condition_number, objective_value) = \
-                            generate_lin_system_from_regression_problem(
-                                n, d, sigma, filepath_in)
+        # To change the bitwidth we need to recompile
+        # the linear regression code, but not oblivc
+        if current_bit_width != bit_width:
+            current_bit_width = bit_width
+            if not RUN_LOCALLY:
+                update_and_compile(ips[0], ips[1],
+                    bit_width=bit_width, compile_oblivc=False)
+                update_and_compile(ips[1], ips[0],
+                    bit_width=bit_width, compile_oblivc=False)
 
-                        logger.info('Wrote instance in file {0}'.format(
-                            filepath_in))
+        logger.info(
+            'Running instance: n={0}, d={1}, sigma={2}, alg={3}, num_iters_cgd={4}, run={5}, bitwidth={6}'.
+            format(
+                n, d, sigma, alg, num_iters_cgd,
+                i + args.instance_num_offset, bit_width))
+        filename_in = 'test_LS_{0}x{1}_{2}_{3}.in'.format(
+            n, d, sigma, i + args.instance_num_offset)
+        filepath_in = os.path.join(
+            dest_folder, filename_in)
 
-                        for party in [1, 2]:
-                            filename_exec = 'test_LS_{0}x{1}_{2}_{3}_{4}_{5}_p{6}.exec'.format(
-                                n, d, sigma, i + args.instance_num_offset, alg, num_iters_cgd, party)
-                            filepath_exec = os.path.join(
-                                dest_folder, filename_exec)
+        (A, b, X, y, lambda_, beta, condition_number, objective_value) = \
+            generate_lin_system_from_regression_problem(
+                n, d, sigma, filepath_in)
 
-                            cmd = '{0} {7} {1} {2} {3} {4} > {5} {6}'.format(
-                                exec_file,
-                                party,
-                                filepath_in,
-                                alg,
-                                num_iters_cgd,
-                                filepath_exec,
-                                '&' if party == 1 else '',
-                                args.port)
+        logger.info('Wrote instance in file {0}'.format(
+            filepath_in))
 
-                            out_filename = os.path.splitext(
-                                filepath_exec)[0] + '.out'
-                            if os.path.exists(out_filename) and not RECOMPUTE:
-                                logger.info('File {} exists. Skipping instance.'.format(out_filename))
+        for party in [1, 2]:
+            filename_exec = 'test_LS_{0}x{1}_{2}_{3}_{4}_{5}_{6}_p{7}.exec'.\
+                format(
+                    n, d, sigma, i + args.instance_num_offset,
+                    alg, bit_width, num_iters_cgd, party)
+            filepath_exec = os.path.join(
+                dest_folder, filename_exec)
 
-                            if RUN_LOCALLY:
-                                out_filename = os.path.splitext(
-                                    filepath_exec)[0] + '.out'
-                                os.system(cmd)
-                                if party == 2:
-                                    parse_output(n, d, X, y, lambda_,
-                                        alg, beta, condition_number,
-                                        objective_value,
-                                        out_filename, filepath_exec)
+            cmd = '{0} {7} {1} {2} {3} {4} > {5} {6}'.format(
+                exec_file,
+                party,
+                filepath_in,
+                alg,
+                num_iters_cgd,
+                filepath_exec,
+                '&' if party == 1 else '',
+                args.port)
 
-                            else:
-                                remote_working_dir = 'secure-distributed-linear-regression/'
-                                run_instance_remotely(
-                                    n, d, X, y, lambda_, alg, beta,
-                                    condition_number,
-                                    objective_value,
-                                    remote_working_dir,
-                                    dest_folder, dest_folder,
-                                    filepath_in, filepath_exec,
-                                    ips[party - 1], cmd, party == 2)
-                                time.sleep(.2)
+            out_filename = os.path.splitext(
+                filepath_exec)[0] + '.out'
+            if os.path.exists(out_filename) and not RECOMPUTE:
+                logger.info('File {} exists. Skipping instance.'.format(out_filename))
 
-                        # Remove instance (they get too big)
-                        logger.info('Deleting file {0}'.format(filepath_in))
-                        #os.system('rm {0}'.format(filepath_in))
+            if RUN_LOCALLY:
+                out_filename = os.path.splitext(
+                    filepath_exec)[0] + '.out'
+                os.system(cmd)
+                if party == 2:
+                    parse_output(n, d, X, y, lambda_,
+                        alg, beta, condition_number,
+                        objective_value,
+                        out_filename, filepath_exec)
+
+            else:
+                remote_working_dir = 'secure-distributed-linear-regression/'
+                run_instance_remotely(
+                    n, d, X, y, lambda_, alg, beta,
+                    condition_number,
+                    objective_value,
+                    remote_working_dir,
+                    dest_folder, dest_folder,
+                    filepath_in, filepath_exec,
+                    ips[party - 1], cmd, party == 2)
+                time.sleep(.2)
+
+        # Remove instance (they get too big)
+        logger.info('Deleting file {0}'.format(filepath_in))
+        os.system('rm {0}'.format(filepath_in))
